Accept fieldname in get_original_value.
Otherwise it depends on having self.fieldname, which is not set here.
Then it falls back to IPrimaryFieldInfo.
In case of a news item with lead image, this does not find anything.

Do not register adapter for IPrimaryFieldInfo in functional tests, only in the integration tests.
I added this registration yesterday to fix some test failures.
I did not see that the test failures were problems that actually needed fixing.
I do that in this commit, plus one in plone.scale.