test_adapters: skip most tests on Plone 5.
For good measure added one test for Plone 5, to see that nothing is returned, which we skip in Plone 6.
Make sure adapters.py can be loaded also on Plone 5, even if it is not used.
The idea is that test_adapters should be importable on Plone 5 and contain at least one test, otherwise the test runner will complain.